// src/state/aether-state-db.generated.d.ts
export interface Projects {
  created_at_ms: number;
  display_name: string;
  id: string;
  origin_url: string | null;
  repo_root: string;
  source: string;
  updated_at_ms: number;
}
// ...
export interface DB {
// ...
  projects: Projects;
// ...
}